+++ pubspec.yaml
@@ -15,7 +15,6 @@ dependencies:
   audioplayers: ^6.1.0
   collection: ^1.17.0
   shared_preferences: ^2.0.15
-  flutter_client_sse: ^2.0.3
 
   english_words: ^4.0.0
   provider: ^6.0.0
